requirements.txt
PySide6>=6.7
SQLAlchemy>=2.0
httpx>=0.27
pydantic>=2.7
tenacity>=8.3
rapidfuzz>=3.9
beautifulsoup4>=4.12
lxml>=5.2
python-dateutil>=2.9

pytest>=8.2
